+++ palletmaster/src/components/Mobile/SkillList_Mobile.tsx
@@ -7,7 +7,8 @@ import TableCell from '@material-ui/core/TableCell';
 import TableHead from '@material-ui/core/TableHead';
 import TableRow from '@material-ui/core/TableRow';
 import TextField from '@material-ui/core/TextField';
-import { skill, character } from '../modules/commonType';
+
+import { skill, character } from '../../modules/commonType';
 
 type Props = {
     skills: skill[];
@@ -52,13 +53,12 @@ const useStyles = makeStyles((theme: Theme) =>
         },
     })
 );
-
 const SkillsTableRow: React.SFC<PropsRow> = (props: PropsRow) => {
     const classes = useStyles();
-    const [skillWork, setSkillWork] = React.useState<number | string>(
+    const [skillWork, setSkillWork] = React.useState<number | undefined>(
         props.row.skillWorkValue
     );
-    const [skillInterest, setSkillInterest] = React.useState<number | string>(
+    const [skillInterest, setSkillInterest] = React.useState<number>(
         props.row.skillInterestValue
     );
 
@@ -86,10 +86,6 @@ const SkillsTableRow: React.SFC<PropsRow> = (props: PropsRow) => {
                     value={skillWork}
                     className={classes.numberInfoField}
                     onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
-                        if (event.target.value === '') {
-                            setSkillWork('');
-                            return;
-                        }
                         if (
                             props.checkSetSkillValue(
                                 props.row.defaultValue +
@@ -131,10 +127,6 @@ const SkillsTableRow: React.SFC<PropsRow> = (props: PropsRow) => {
                     value={skillInterest}
                     className={classes.numberInfoField}
                     onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
-                        if (event.target.value === '') {
-                            setSkillWork('');
-                            return;
-                        }
                         if (
                             props.checkSetSkillValue(
                                 props.row.defaultValue +
@@ -173,7 +165,7 @@ const SkillsTableRow: React.SFC<PropsRow> = (props: PropsRow) => {
     );
 };
 
-const SimpleList: React.SFC<Props> = (props: Props) => {
+export default function SimpleList(props: Props) {
     const classes = useStyles();
 
     return (
@@ -190,6 +182,8 @@ const SimpleList: React.SFC<Props> = (props: Props) => {
                         <TableCell className={classes.number} align="right">
                             Type
                         </TableCell>
+                    </TableRow>
+                    <TableRow>
                         <TableCell className={classes.number} align="right">
                             Work
                         </TableCell>
@@ -229,6 +223,4 @@ const SimpleList: React.SFC<Props> = (props: Props) => {
             </Table>
         </div>
     );
-};
-
-export default SimpleList;
+}
